+++ src/main/java/com/example/ApartmentRenovationCostEstimate/repositories/ProductRepository.java
@@ -3,5 +3,5 @@ package com.example.ApartmentRenovationCostEstimate.repositories;
 import com.example.ApartmentRenovationCostEstimate.entity.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Product, Integer> {
+public interface ProductRepository extends CrudRepository<Product, Long> {
 }
